use more skalibs, reduce compiler warnings

+++ src/socklog/taia.h
@@ -7,8 +7,8 @@
 
 struct taia {
   tai_t sec;
-  unsigned long nano; /* 0...999999999 */
-  unsigned long atto; /* 0...999999999 */
+  uint32_t nano; /* 0...999999999 */
+  uint32_t atto; /* 0...999999999 */
 } ;
 
 extern void taia_tai(const struct taia *,tai_t *);
